fix bugs for supporting single class

+++ layers/modules/multibox_loss.py
@@ -6,6 +6,7 @@ from torch.autograd import Variable
 from ..box_utils import match, log_sum_exp
 from data import ZH as cfg
 
+
 class MultiBoxLoss(nn.Module):
     """SSD Weighted Loss Function
     Compute Targets:
@@ -46,16 +47,15 @@ class MultiBoxLoss(nn.Module):
 
     def forward(self, predictions, targets):
         """Multibox Loss
-        Args:
-            predictions (tuple): A tuple containing loc preds, conf preds,
-            and prior boxes from SSD net.
-                conf shape: torch.size(batch_size,num_priors,num_classes)
-                loc shape: torch.size(batch_size,num_priors,4)
-                priors shape: torch.size(num_priors,4)
-
-            targets (tensor): Ground truth boxes and labels for a batch,
-                shape: [batch_size,num_objs,5] (last idx is the label).
-        """
+              Args:
+                  predictions (tuple): A tuple containing loc preds, conf preds,
+                  and prior boxes from SSD net.
+                      conf shape: torch.size(batch_size,num_priors,num_classes)
+                      loc shape: torch.size(batch_size,num_priors,4)
+                      priors shape: torch.size(num_priors,4)
+                  targets (tensor): Ground truth boxes and labels for a batch,
+                      shape: [batch_size,num_objs,5] (last idx is the label).
+              """
         loc_data, conf_data, priors = predictions
         num = loc_data.size(0)
         priors = priors[:loc_data.size(1), :]
@@ -98,19 +98,19 @@ class MultiBoxLoss(nn.Module):
         _, loss_idx = loss_c.sort(1, descending=True)
         _, idx_rank = loss_idx.sort(1)
         num_pos = pos.long().sum(1, keepdim=True)
-        num_neg = torch.clamp(self.negpos_ratio*num_pos, max=pos.size(1)-1)
+        num_neg = torch.clamp(self.negpos_ratio * num_pos, max=pos.size(1) - 1)
         neg = idx_rank < num_neg.expand_as(idx_rank)
 
         # Confidence Loss Including Positive and Negative Examples
         pos_idx = pos.unsqueeze(2).expand_as(conf_data)
         neg_idx = neg.unsqueeze(2).expand_as(conf_data)
-        conf_p = conf_data[(pos_idx+neg_idx).gt(0)].view(-1, self.num_classes)
-        targets_weighted = conf_t[(pos+neg).gt(0)]
+        conf_p = conf_data[(pos_idx + neg_idx).gt(0)].view(-1, self.num_classes)
+        targets_weighted = conf_t[(pos + neg).gt(0)]
         loss_c = F.cross_entropy(conf_p, targets_weighted, size_average=False)
 
         # Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
 
-        N = num_pos.data.sum().float()
-        loss_l /= N
-        loss_c /= N
+        N = num_pos.data.sum()
+        loss_l /= N.float()
+        loss_c /= N.float()
         return loss_l, loss_c
